Fix several clippy warnings
Fix several clippy warnings related to:
- Boolean to `int` conversion using `if`.
- Borrowed expression implements the required traits.
- Unnecessary castings.

In the following crates:
- `collections`
- `lib`
- `nano-zkp`
- `network-libp2p`
- `trie` primitives
- `rpc-server`
- `tools`
- `utils`

+++ primitives/trie/src/key_nibbles.rs
@@ -326,7 +326,7 @@ impl Serialize for KeyNibbles {
     fn serialize<W: WriteBytesExt>(&self, writer: &mut W) -> Result<usize, SerializingError> {
         let mut size = 2;
         writer.write_u8(self.length)?;
-        writer.write_u8(self.bytes_length as u8)?;
+        writer.write_u8(self.bytes_length)?;
         size += writer.write(&self.bytes[..self.bytes_length as usize])?;
         Ok(size)
     }
